+++ macholib/dylib.py
@@ -6,7 +6,7 @@ import re
 
 __all__ = ['dylib_info']
 
-DYLIB_RE = re.compile(r"""(?x)
+_DYLIB_RE = re.compile(r"""(?x)
 (?P<location>^.*)(?:^|/)
 (?P<name>
     (?P<shortname>\w+?)
@@ -36,7 +36,7 @@ def dylib_info(filename):
     Note that SomeVersion and Suffix are optional and may be None
     if not present.
     """
-    is_dylib = DYLIB_RE.match(filename)
+    is_dylib = _DYLIB_RE.match(filename)
     if not is_dylib:
         return None
     return is_dylib.groupdict()