+++ macholib/itergraphreport.py
@@ -4,8 +4,8 @@ Utilities for creating dot output from a MachOGraph
 XXX: need to rewrite this based on altgraph.Dot
 """
 
-from pkg_resources import require
-require("altgraph")
+from collections import deque
+from itertools import imap
 
 __all__ = ['itergraphreport']