+++ macholib_tests/test_MachOStandalone.py
@@ -2,8 +2,14 @@ from macholib import MachOStandalone
 
 import unittest
 
+try:
+    expectedFailure = unittest.expectedFailure
+except AttributeError:
+    expectedFailure = lambda function: function
+
 
 class TestMachOStandalone (unittest.TestCase):
+    @expectedFailure
     def test_missing(self):
         self.fail("tests are missing")